Fixed compatibility issues, added configuration to rescale labels

+++ mmdet/core/bbox/bbox_target.py
@@ -43,7 +43,7 @@ def bbox_target_single(pos_bboxes,
     num_pos = pos_bboxes.size(0)
     num_neg = neg_bboxes.size(0)
     num_samples = num_pos + num_neg
-    labels = pos_bboxes.new_zeros(num_samples, dtype=torch.long)
+    labels = pos_bboxes.new_zeros(num_samples, dtype=torch.double)
     label_weights = pos_bboxes.new_zeros(num_samples)
     bbox_targets = pos_bboxes.new_zeros(num_samples, 4)
     bbox_weights = pos_bboxes.new_zeros(num_samples, 4)
